+++ semgrep-core/src/synthesizing/Pattern_from_Code.ml
@@ -31,8 +31,12 @@ exception UnexpectedCase of string
 type named_variants =
   (string * Pattern.t) list
 
-type environment = { count : int; mapping : (expr * expr) list;
-                     has_type : bool }
+type environment = {
+  config: Config_semgrep.t;
+  count: int;
+  mapping: (expr * expr) list;
+  has_type: bool;
+}
 
 (*****************************************************************************)
 (* State helpers *)
@@ -44,7 +48,9 @@ let lookup env e =
   let rec look = function
     | [] -> None
     | (e1, e2)::xs ->
-        if Matching_generic.equal_ast_binded_code (E e) (E e1) then Some e2 else look xs
+        if Matching_generic.equal_ast_binded_code env.config (E e) (E e1)
+        then Some e2
+        else look xs
   in
   look mapping
 
@@ -101,7 +107,13 @@ let get_id ?(with_type=false) env e =
            | _ -> (notype_id, has_type)
           )
       in
-      ({ count = env.count + 1; mapping = (e, new_id)::(env.mapping); has_type = new_has_type }, new_id)
+      let env' =
+        { env with
+          count = env.count + 1;
+          mapping = (e, new_id)::(env.mapping);
+          has_type = new_has_type;
+        } in
+      (env', new_id)
 
 let has_nested_call = List.find_opt (fun x -> match x with Arg(Call _) -> true | _ -> false)
 
@@ -328,13 +340,13 @@ and generalize_any a env =
   | S s -> generalize_stmt s env
   | _ -> []
 
-let generalize a =
-  (generalize_any a { count = 1; mapping = []; has_type = false })
+let generalize config a =
+  (generalize_any a { config; count = 1; mapping = []; has_type = false })
 
 (*****************************************************************************)
 (* Entry point *)
 (*****************************************************************************)
 
-let from_any a =
+let from_any config a =
   ("exact match", a)::
-  generalize a
+  generalize config a